requirements.txt
pandas
adal
requests
uuid
python-dotenv
numpy
wget